backend/requirements.txt
# Web
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20
pydantic[email]==2.9.0
pydantic-settings==2.6.0
email-validator==2.2.0

# DB
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Cache / queue
redis[hiredis]==5.2.0

# LLM / embeddings
anthropic==0.40.0
openai==1.55.0
sentence-transformers==3.3.1
# torch CPU-only is enough for embedding inference at our scale.
torch==2.5.1

# KB ingestion
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7
beautifulsoup4==4.12.3
tiktoken==0.8.0
pyyaml==6.0.2

# Security / utils
python-jose[cryptography]==3.3.0
httpx==0.28.0
tenacity==9.0.0
structlog==24.4.0
python-dotenv==1.0.1

# Polish text
regex==2024.11.6

# Dev
pytest==8.3.0
pytest-asyncio==0.25.0
ruff==0.8.0
